pyproject.toml
[tool.poetry]
name = "Renewable Data Rangers"
version = "0.2.0"
description = "Data engineer project for ISEP seminar"
authors = ["Tiago Reis <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "^3.5.1"
numpy = "^1.21.0"
pandas = "2.0.2"
psycopg2-binary = "2.9.6"
SQLAlchemy = "2.0.17"
faker = "20.1.0"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
boto3 = "^1.34.113"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"